build: use `npm install` to work around npm/cli#558 (#6571)
the lockfile we use was generated on a mac & includes
`fsevents` as a dependency. while this is optional for
OSX it's not available for linux & this causes problems
with the `npm ci` command.  npm v7 will correct this
but until we're using that we should switch to
using npm install