fix: Only check for discriminator conflicts on STI entities (#2985)
If an entity has the same discriminator value as one using `TableInheritance` this should be allowed as 
there is no way for them to conflict.

This commit also fixes the discriminatorColumn 
check to check all the entities it should be
checking. Only the parent in an STI setup has the
"STI" inheritancePattern, the children just have the 
"entity-child" tableType. Before two children with 
the same discriminatorValue would be
(incorrectly) allowed as they'd never even get 
checked.

Fixes #2984