fix: oracle cannot support DB in table identifiers (#7954)
oracle queries cannot support the identifier including the database in this way (`XE.SYSTEM.table`)

Instead, if you should specify a separate database a link must be set up and then
the database link identifier is included after the table. (`SYSTEM.table@XE`)

However, because that's rather difficult to set up lets not for now & just prevent
inclusion of the database in the case that we fetch it.